requirements.txt
absl-py==1.2.0
astunparse==1.6.3
beautifulsoup4==4.11.1
bidict==0.22.0
cachetools==5.2.0
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
charset-normalizer==2.1.0
click @ file:///C:/ci/click_1646056762388/work
colorama @ file:///C:/Windows/TEMP/abs_9439aeb1-0254-449a-96f7-33ab5eb17fc8apleb4yn/croots/recipe/colorama_1657009099097/work
deepface==0.0.75
filelock==3.8.0
fire==0.4.0
Flask @ file:///C:/ci/flask_1658860024433/work
flatbuffers==1.12
gast==0.4.0
gdown==4.5.1
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==3.3
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
mailer==0.8.1
Markdown==3.4.1
MarkupSafe @ file:///C:/ci/markupsafe_1654508036328/work
mtcnn==0.1.1
neo4j==4.4.5
numpy==1.23.1
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.1
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
retina-face==0.0.12
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug @ file:///opt/conda/conda-bld/werkzeug_1645628268370/work
wincertstore==0.2
wrapt==1.14.1
